pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptproxy"
version = "0.1.0"
description = "A local-first, developer-facing intercepting prompt proxy for LLM traffic"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "spacy>=3.7.0",
    "litellm>=1.35.0",
    "rich>=13.7.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "python-json-logger>=2.0.0",
]

[project.scripts]
promptproxy = "promptproxy.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["promptproxy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"